test: Move an integration test of hot compact_one_partition to lib
