+++ compactor/src/hot.rs
@@ -206,21 +206,16 @@ fn query_times(time_provider: Arc<dyn TimeProvider>) -> Vec<(u64, Timestamp)> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{
-        compact::Compactor, compact_one_partition, handler::CompactorConfig,
-        parquet_file_filtering, parquet_file_lookup, ParquetFilesForCompaction,
-    };
-    use arrow_util::assert_batches_sorted_eq;
+    use crate::{compact::Compactor, handler::CompactorConfig};
     use backoff::BackoffConfig;
     use data_types::{
-        ColumnId, ColumnSet, ColumnType, CompactionLevel, ParquetFileId, ParquetFileParams,
-        SequenceNumber, ShardIndex,
+        ColumnId, ColumnSet, CompactionLevel, ParquetFileParams, SequenceNumber, ShardIndex,
     };
     use iox_query::exec::Executor;
-    use iox_tests::util::{TestCatalog, TestParquetFileBuilder};
-    use iox_time::{SystemProvider, TimeProvider};
+    use iox_tests::util::TestCatalog;
+    use iox_time::SystemProvider;
     use parquet_file::storage::ParquetStorage;
-    use std::{collections::HashMap, sync::Arc};
+    use std::sync::Arc;
     use uuid::Uuid;
 
     fn make_compactor_config() -> CompactorConfig {
@@ -493,278 +488,5 @@ mod tests {
         assert_eq!(*candidates[1].table, another_table);
         assert_eq!(candidates[1].partition_key, another_partition.partition_key);
         assert_eq!(candidates[1].sort_key, another_partition.sort_key()); // this sort key is Some(tag1, time)
-    }
-
-    // A quite sophisticated integration test
-    // Beside lp data, every value min/max sequence numbers and min/max time are important
-    // to have a combination of needed tests in this test function
-    #[tokio::test]
-    async fn test_compact_hot_partition_many_files() {
-        test_helpers::maybe_start_logging();
-        let catalog = TestCatalog::new();
-
-        // lp1 does not overlap with any other level 0
-        let lp1 = vec![
-            "table,tag1=WA field_int=1000i 10",
-            "table,tag1=VT field_int=10i 20",
-        ]
-        .join("\n");
-
-        // lp2 overlaps with lp3
-        let lp2 = vec![
-            "table,tag1=WA field_int=1000i 8000", // will be eliminated due to duplicate
-            "table,tag1=VT field_int=10i 10000",
-            "table,tag1=UT field_int=70i 20000",
-        ]
-        .join("\n");
-
-        // lp3 overlaps with lp2
-        let lp3 = vec![
-            "table,tag1=WA field_int=1500i 8000", // latest duplicate and kept
-            "table,tag1=VT field_int=10i 6000",
-            "table,tag1=UT field_int=270i 25000",
-        ]
-        .join("\n");
-
-        // lp4 does not overlap with any
-        let lp4 = vec![
-            "table,tag2=WA,tag3=10 field_int=1600i 28000",
-            "table,tag2=VT,tag3=20 field_int=20i 26000",
-        ]
-        .join("\n");
-
-        // lp5 overlaps with lp1
-        let lp5 = vec![
-            "table,tag2=PA,tag3=15 field_int=1601i 9",
-            "table,tag2=OH,tag3=21 field_int=21i 25",
-        ]
-        .join("\n");
-
-        // lp6 does not overlap with any
-        let lp6 = vec![
-            "table,tag2=PA,tag3=15 field_int=81601i 90000",
-            "table,tag2=OH,tag3=21 field_int=421i 91000",
-        ]
-        .join("\n");
-
-        let ns = catalog.create_namespace("ns").await;
-        let shard = ns.create_shard(1).await;
-        let table = ns.create_table("table").await;
-        table.create_column("field_int", ColumnType::I64).await;
-        table.create_column("tag1", ColumnType::Tag).await;
-        table.create_column("tag2", ColumnType::Tag).await;
-        table.create_column("tag3", ColumnType::Tag).await;
-        table.create_column("time", ColumnType::Time).await;
-
-        let partition = table.with_shard(&shard).create_partition("part").await;
-        let time = Arc::new(SystemProvider::new());
-        let config = CompactorConfig {
-            max_desired_file_size_bytes: 10_000,
-            percentage_max_file_size: 30,
-            split_percentage: 80,
-            max_number_partitions_per_shard: 1,
-            min_number_recent_ingested_files_per_partition: 1,
-            hot_multiple: 4,
-            memory_budget_bytes: 100_000_000,
-        };
-
-        let metrics = Arc::new(metric::Registry::new());
-        let compactor = Arc::new(Compactor::new(
-            vec![shard.shard.id],
-            Arc::clone(&catalog.catalog),
-            ParquetStorage::new(Arc::clone(&catalog.object_store)),
-            Arc::new(Executor::new(1)),
-            Arc::new(SystemProvider::new()),
-            BackoffConfig::default(),
-            config,
-            Arc::clone(&metrics),
-        ));
-
-        // parquet files that are all in the same partition
-        let mut size_overrides = HashMap::<ParquetFileId, i64>::default();
-
-        // pf1 does not overlap with any other level 0
-        let builder = TestParquetFileBuilder::default()
-            .with_line_protocol(&lp1)
-            .with_max_seq(3)
-            .with_min_time(10)
-            .with_max_time(20)
-            .with_creation_time(time.now());
-        let f = partition.create_parquet_file(builder).await;
-        size_overrides.insert(
-            f.parquet_file.id,
-            compactor.config.max_desired_file_size_bytes as i64 + 10,
-        );
-
-        // pf2 overlaps with pf3
-        let builder = TestParquetFileBuilder::default()
-            .with_line_protocol(&lp2)
-            .with_max_seq(5)
-            .with_min_time(8_000)
-            .with_max_time(20_000)
-            .with_creation_time(time.now());
-        let f = partition.create_parquet_file(builder).await;
-        size_overrides.insert(f.parquet_file.id, 100); // small file
-
-        // pf3 overlaps with pf2
-        let builder = TestParquetFileBuilder::default()
-            .with_line_protocol(&lp3)
-            .with_max_seq(10)
-            .with_min_time(6_000)
-            .with_max_time(25_000)
-            .with_creation_time(time.now());
-        let f = partition.create_parquet_file(builder).await;
-        size_overrides.insert(f.parquet_file.id, 100); // small file
-
-        // pf4 does not overlap with any but is small
-        let builder = TestParquetFileBuilder::default()
-            .with_line_protocol(&lp4)
-            .with_max_seq(18)
-            .with_min_time(26_000)
-            .with_max_time(28_000)
-            .with_creation_time(time.now());
-        let f = partition.create_parquet_file(builder).await;
-        size_overrides.insert(f.parquet_file.id, 100); // small file
-
-        // pf5 was created in a previous compaction cycle; overlaps with pf1
-        let builder = TestParquetFileBuilder::default()
-            .with_line_protocol(&lp5)
-            .with_max_seq(1)
-            .with_min_time(9)
-            .with_max_time(25)
-            .with_creation_time(time.now())
-            .with_compaction_level(CompactionLevel::FileNonOverlapped);
-        let f = partition.create_parquet_file(builder).await;
-        size_overrides.insert(f.parquet_file.id, 100); // small file
-
-        // pf6 was created in a previous compaction cycle; does not overlap with any
-        let builder = TestParquetFileBuilder::default()
-            .with_line_protocol(&lp6)
-            .with_max_seq(20)
-            .with_min_time(90000)
-            .with_max_time(91000)
-            .with_creation_time(time.now())
-            .with_compaction_level(CompactionLevel::FileNonOverlapped);
-        let f = partition.create_parquet_file(builder).await;
-        size_overrides.insert(f.parquet_file.id, 100); // small file
-
-        // should have 4 level-0 files before compacting
-        let count = catalog.count_level_0_files(shard.shard.id).await;
-        assert_eq!(count, 4);
-
-        let query_times = query_times(compactor.catalog.time_provider());
-
-        // ------------------------------------------------
-        // Compact
-        let partition_candidates = hot_partitions_for_shard(
-            Arc::clone(&compactor.catalog),
-            shard.shard.id,
-            &query_times,
-            1,
-            1,
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(partition_candidates.len(), 1);
-
-        let table_columns = compactor
-            .table_columns(&partition_candidates)
-            .await
-            .unwrap();
-        let mut partition_candidates = compactor
-            .add_info_to_partitions(&partition_candidates, &table_columns)
-            .await
-            .unwrap();
-        let partition = partition_candidates.pop().unwrap();
-
-        let parquet_files_for_compaction =
-            parquet_file_lookup::ParquetFilesForCompaction::for_partition_with_size_overrides(
-                Arc::clone(&compactor.catalog),
-                Arc::clone(&partition),
-                &size_overrides,
-            )
-            .await
-            .unwrap();
-
-        let ParquetFilesForCompaction {
-            level_0,
-            level_1,
-            .. // Ignore other levels
-        } = parquet_files_for_compaction;
-
-        let to_compact = parquet_file_filtering::filter_parquet_files(
-            partition,
-            level_0,
-            level_1,
-            compactor.config.memory_budget_bytes,
-            &compactor.parquet_file_candidate_gauge,
-            &compactor.parquet_file_candidate_bytes,
-        );
-
-        let to_compact = to_compact.into();
-
-        compact_one_partition(&compactor, to_compact, "hot", true)
-            .await
-            .unwrap();
-
-        // Should have 3 non-soft-deleted files:
-        //
-        // - the level 1 file that didn't overlap with anything
-        // - the two newly created after compacting and splitting pf1, pf2, pf3, pf4, pf5
-        let mut files = catalog.list_by_table_not_to_delete(table.table.id).await;
-        assert_eq!(files.len(), 3);
-        let files_and_levels: Vec<_> = files
-            .iter()
-            .map(|f| (f.id.get(), f.compaction_level))
-            .collect();
-        assert_eq!(
-            files_and_levels,
-            vec![
-                (6, CompactionLevel::FileNonOverlapped),
-                (7, CompactionLevel::FileNonOverlapped),
-                (8, CompactionLevel::FileNonOverlapped),
-            ]
-        );
-
-        // ------------------------------------------------
-        // Verify the parquet file content
-
-        // Later compacted file
-        let file1 = files.pop().unwrap();
-        let batches = table.read_parquet_file(file1).await;
-        assert_batches_sorted_eq!(
-            &[
-                "+-----------+------+------+------+-----------------------------+",
-                "| field_int | tag1 | tag2 | tag3 | time                        |",
-                "+-----------+------+------+------+-----------------------------+",
-                "| 1600      |      | WA   | 10   | 1970-01-01T00:00:00.000028Z |",
-                "| 20        |      | VT   | 20   | 1970-01-01T00:00:00.000026Z |",
-                "| 270       | UT   |      |      | 1970-01-01T00:00:00.000025Z |",
-                "+-----------+------+------+------+-----------------------------+",
-            ],
-            &batches
-        );
-
-        // Earlier compacted file
-        let file0 = files.pop().unwrap();
-        let batches = table.read_parquet_file(file0).await;
-        assert_batches_sorted_eq!(
-            &[
-                "+-----------+------+------+------+--------------------------------+",
-                "| field_int | tag1 | tag2 | tag3 | time                           |",
-                "+-----------+------+------+------+--------------------------------+",
-                "| 10        | VT   |      |      | 1970-01-01T00:00:00.000000020Z |",
-                "| 10        | VT   |      |      | 1970-01-01T00:00:00.000006Z    |",
-                "| 10        | VT   |      |      | 1970-01-01T00:00:00.000010Z    |",
-                "| 1000      | WA   |      |      | 1970-01-01T00:00:00.000000010Z |",
-                "| 1500      | WA   |      |      | 1970-01-01T00:00:00.000008Z    |",
-                "| 1601      |      | PA   | 15   | 1970-01-01T00:00:00.000000009Z |",
-                "| 21        |      | OH   | 21   | 1970-01-01T00:00:00.000000025Z |",
-                "| 70        | UT   |      |      | 1970-01-01T00:00:00.000020Z    |",
-                "+-----------+------+------+------+--------------------------------+",
-            ],
-            &batches
-        );
     }
 }
